+++ src/Narvalo.Brouillons/Edu/Linq/Bananas.cs
@@ -6,6 +6,7 @@ namespace Narvalo.Edu.Linq
     using System.Collections.Generic;
 
     using Narvalo.Fx;
+    using Narvalo.Fx.Linq;
 
     // LINQ from scratch.
     public static class Bananas