+++ src/Narvalo.Brouillons/Edu/Monads/Samples/Monad.g.cs
@@ -425,8 +425,7 @@ namespace Narvalo.Edu.Monads.Samples
     } // End of the class Monad.
 }
 
-// Extensions for core CLR types are put into a separate namespace.
-namespace Narvalo.Edu.Monads.Samples.Advanced
+namespace Narvalo.Edu.Monads.Samples
 {
     using System;
     using System.Diagnostics.Contracts;