+++ src/Narvalo.Core/Fx/Linq/Maybe.g.cs
@@ -29,56 +29,72 @@ using global::System.Diagnostics.CodeAnalysis;
 [module: SuppressMessage("StyleCop.CSharp.LayoutRules", "SA1507:CodeMustNotContainMultipleBlankLinesInARow",
     Justification = "[GeneratedCode] Newline rules are disabled for T4 templates.")]
 
-namespace Narvalo.Fx.Advanced
+namespace Narvalo.Fx.Linq
 {
     using System;
     using System.Collections.Generic;
     using System.Diagnostics.Contracts;
 
     using Narvalo.Fx;
-    using Narvalo.Fx.Advanced.Internal;
+    using Narvalo.Fx.Linq.Internal;
 
     /// <summary>
-    /// Provides extension methods for <see cref="IEnumerable{T}"/> that depend on the <see cref="Output{T}"/> class..
+    /// Provides extension methods for <see cref="IEnumerable{T}"/> that depend on the <see cref="Maybe{T}"/> class..
     /// </summary>
     [global::System.CodeDom.Compiler.GeneratedCode("Microsoft.VisualStudio.TextTemplating.12.0", "12.0.0.0")]
     [global::System.Diagnostics.DebuggerNonUserCode]
-    public static partial class EnumerableOutputExtensions
+    public static partial class EnumerableMaybeExtensions
     {
         #region Basic Monad functions (Prelude)
 
         /// <remarks>
         /// Named <c>sequence</c> in Haskell parlance.
         /// </remarks>
-        public static Output<IEnumerable<TSource>> Collect<TSource>(
-            this IEnumerable<Output<TSource>> @this)
+        public static Maybe<IEnumerable<TSource>> Collect<TSource>(
+            this IEnumerable<Maybe<TSource>> @this)
         {
             Acknowledge.Object(@this);
-            Contract.Ensures(Contract.Result<Output<IEnumerable<TSource>>>() != null);
+            Contract.Ensures(Contract.Result<Maybe<IEnumerable<TSource>>>() != null);
 
             return @this.CollectCore();
         }
 
         #endregion
-    } // End of the class EnumerableOutputExtensions.
+
+        #region Generalisations of list functions (Prelude)
+
+        /// <remarks>
+        /// Named <c>msum</c> in Haskell parlance.
+        /// </remarks>
+        public static Maybe<TSource> Sum<TSource>(
+            this IEnumerable<Maybe<TSource>> @this)
+        {
+            Acknowledge.Object(@this);
+            Contract.Ensures(Contract.Result<Maybe<TSource>>() != null);
+
+            return @this.SumCore();
+        }
+
+        #endregion
+    } // End of the class EnumerableMaybeExtensions.
 
     /// <content>
     /// Provides extension methods for <see cref="IEnumerable{T}"/>.
     /// </content>
-    public static partial class EnumerableOutputExtensions
+    public static partial class EnumerableMaybeExtensions
     {
         #region Basic Monad functions (Prelude)
 
         /// <remarks>
         /// Named <c>mapM</c> in Haskell parlance.
         /// </remarks>
-        public static Output<IEnumerable<TResult>> Map<TSource, TResult>(
+        public static Maybe<IEnumerable<TResult>> Map<TSource, TResult>(
             this IEnumerable<TSource> @this,
-            Func<TSource, Output<TResult>> funM)
+            Func<TSource, Maybe<TResult>> funM)
         {
             Acknowledge.Object(@this);
             Contract.Requires(funM != null);
-            Contract.Ensures(Contract.Result<Output<IEnumerable<TResult>>>() != null);
+            Contract.Ensures(Contract.Result<Maybe<IEnumerable<TResult>>>() != null);
 
             return @this.MapCore(funM);
         }
@@ -93,7 +109,7 @@ namespace Narvalo.Fx.Advanced
         /// </remarks>
         public static IEnumerable<TSource> Filter<TSource>(
             this IEnumerable<TSource> @this,
-            Func<TSource, Output<bool>> predicateM)
+            Func<TSource, Maybe<bool>> predicateM)
         {
             Acknowledge.Object(@this);
             Contract.Requires(predicateM != null);
@@ -105,13 +121,14 @@ namespace Narvalo.Fx.Advanced
         /// <remarks>
         /// Named <c>mapAndUnzipM</c> in Haskell parlance.
         /// </remarks>
-        public static Output<Tuple<IEnumerable<TFirst>, IEnumerable<TSecond>>>
+        public static Maybe<Tuple<IEnumerable<TFirst>, IEnumerable<TSecond>>>
             MapAndUnzip<TSource, TFirst, TSecond>(
             this IEnumerable<TSource> @this,
-            Func<TSource, Output<Tuple<TFirst, TSecond>>> funM)
+            Func<TSource, Maybe<Tuple<TFirst, TSecond>>> funM)
         {
             Acknowledge.Object(@this);
             Contract.Requires(funM != null);
+            Contract.Ensures(Contract.Result<Maybe<Tuple<IEnumerable<TFirst>, IEnumerable<TSecond>>>>() != null);
 
             return @this.MapAndUnzipCore(funM);
         }
@@ -119,15 +136,15 @@ namespace Narvalo.Fx.Advanced
         /// <remarks>
         /// Named <c>zipWithM</c> in Haskell parlance.
         /// </remarks>
-        public static Output<IEnumerable<TResult>> Zip<TFirst, TSecond, TResult>(
+        public static Maybe<IEnumerable<TResult>> Zip<TFirst, TSecond, TResult>(
             this IEnumerable<TFirst> @this,
             IEnumerable<TSecond> second,
-            Func<TFirst, TSecond, Output<TResult>> resultSelectorM)
+            Func<TFirst, TSecond, Maybe<TResult>> resultSelectorM)
         {
             Acknowledge.Object(@this);
             Contract.Requires(second != null);
             Contract.Requires(resultSelectorM != null);
-            Contract.Ensures(Contract.Result<Output<IEnumerable<TResult>>>() != null);
+            Contract.Ensures(Contract.Result<Maybe<IEnumerable<TResult>>>() != null);
 
             return @this.ZipCore(second, resultSelectorM);
         }
@@ -135,13 +152,14 @@ namespace Narvalo.Fx.Advanced
         /// <remarks>
         /// Named <c>foldM</c> in Haskell parlance.
         /// </remarks>
-        public static Output<TAccumulate> Fold<TSource, TAccumulate>(
+        public static Maybe<TAccumulate> Fold<TSource, TAccumulate>(
             this IEnumerable<TSource> @this,
             TAccumulate seed,
-            Func<TAccumulate, TSource, Output<TAccumulate>> accumulatorM)
+            Func<TAccumulate, TSource, Maybe<TAccumulate>> accumulatorM)
         {
             Acknowledge.Object(@this);
             Contract.Requires(accumulatorM != null);
+            Contract.Ensures(Contract.Result<Maybe<TAccumulate>>() != null);
 
             return @this.FoldCore(seed, accumulatorM);
         }
@@ -150,33 +168,36 @@ namespace Narvalo.Fx.Advanced
 
         #region Aggregate Operators
         
-        public static Output<TAccumulate> FoldBack<TSource, TAccumulate>(
+        public static Maybe<TAccumulate> FoldBack<TSource, TAccumulate>(
             this IEnumerable<TSource> @this,
             TAccumulate seed,
-            Func<TAccumulate, TSource, Output<TAccumulate>> accumulatorM)
+            Func<TAccumulate, TSource, Maybe<TAccumulate>> accumulatorM)
         {
             Acknowledge.Object(@this);
             Contract.Requires(accumulatorM != null);
+            Contract.Ensures(Contract.Result<Maybe<TAccumulate>>() != null);
 
             return @this.FoldBackCore(seed, accumulatorM);
         }
         
-        public static Output<TSource> Reduce<TSource>(
+        public static Maybe<TSource> Reduce<TSource>(
             this IEnumerable<TSource> @this,
-            Func<TSource, TSource, Output<TSource>> accumulatorM)
+            Func<TSource, TSource, Maybe<TSource>> accumulatorM)
         {
             Acknowledge.Object(@this);
             Contract.Requires(accumulatorM != null);
+            Contract.Ensures(Contract.Result<Maybe<TSource>>() != null);
 
             return @this.ReduceCore(accumulatorM);
         }
         
-        public static Output<TSource> ReduceBack<TSource>(
+        public static Maybe<TSource> ReduceBack<TSource>(
             this IEnumerable<TSource> @this,
-            Func<TSource, TSource, Output<TSource>> accumulatorM)
+            Func<TSource, TSource, Maybe<TSource>> accumulatorM)
         {
             Acknowledge.Object(@this);
             Contract.Requires(accumulatorM != null);
+            Contract.Ensures(Contract.Result<Maybe<TSource>>() != null);
 
             return @this.ReduceBackCore(accumulatorM);
         }
@@ -188,15 +209,16 @@ namespace Narvalo.Fx.Advanced
         /// <remarks>
         /// <para>Haskell use a different signature.</para>
         /// </remarks>
-        public static Output<TAccumulate> Fold<TSource, TAccumulate>(
+        public static Maybe<TAccumulate> Fold<TSource, TAccumulate>(
             this IEnumerable<TSource> @this,
             TAccumulate seed,
-            Func<TAccumulate, TSource, Output<TAccumulate>> accumulatorM,
-            Func<Output<TAccumulate>, bool> predicate)
+            Func<TAccumulate, TSource, Maybe<TAccumulate>> accumulatorM,
+            Func<Maybe<TAccumulate>, bool> predicate)
         {
             Acknowledge.Object(@this);
             Contract.Requires(accumulatorM != null);
             Contract.Requires(predicate != null);
+            Contract.Ensures(Contract.Result<Maybe<TAccumulate>>() != null);
 
             return @this.FoldCore(seed, accumulatorM, predicate);
         }
@@ -204,23 +226,24 @@ namespace Narvalo.Fx.Advanced
         /// <remarks>
         /// <para>Haskell use a different signature.</para>
         /// </remarks>
-        public static Output<TSource> Reduce<TSource>(
+        public static Maybe<TSource> Reduce<TSource>(
             this IEnumerable<TSource> @this,
-            Func<TSource, TSource, Output<TSource>> accumulatorM,
-            Func<Output<TSource>, bool> predicate)
+            Func<TSource, TSource, Maybe<TSource>> accumulatorM,
+            Func<Maybe<TSource>, bool> predicate)
         {
             Acknowledge.Object(@this);
             Contract.Requires(accumulatorM != null);
             Contract.Requires(predicate != null);
+            Contract.Ensures(Contract.Result<Maybe<TSource>>() != null);
 
             return @this.ReduceCore(accumulatorM, predicate);
         }
 
         #endregion
-    } // End of the class EnumerableOutputExtensions.
+    } // End of the class EnumerableMaybeExtensions.
 }
 
-namespace Narvalo.Fx.Advanced.Internal
+namespace Narvalo.Fx.Linq.Internal
 {
     using System;
     using System.Collections.Generic;
@@ -232,50 +255,61 @@ namespace Narvalo.Fx.Advanced.Internal
     using Narvalo.Fx;
 
     /// <summary>
-    /// Provides the core extension methods for <see cref="IEnumerable{T}"/> that depend on the <see cref="Output{T}"/> class.
+    /// Provides the core extension methods for <see cref="IEnumerable{T}"/> that depend on the <see cref="Maybe{T}"/> class.
     /// </summary>
-    internal static partial class EnumerableOutputExtensions
+    internal static partial class EnumerableMaybeExtensions
     {
         [SuppressMessage("Microsoft.Performance", "CA1811:AvoidUncalledPrivateCode",
             Justification = "[GeneratedCode] This method has been overridden locally.")]
-        internal static Output<IEnumerable<TSource>> CollectCore<TSource>(
-            this IEnumerable<Output<TSource>> @this)
+        internal static Maybe<IEnumerable<TSource>> CollectCore<TSource>(
+            this IEnumerable<Maybe<TSource>> @this)
         {
             Acknowledge.Object(@this);
-            Contract.Ensures(Contract.Result<Output<IEnumerable<TSource>>>() != null);
+            Contract.Ensures(Contract.Result<Maybe<IEnumerable<TSource>>>() != null);
 
-            var seed = Output.Success(Enumerable.Empty<TSource>());
-            Func<Output<IEnumerable<TSource>>, Output<TSource>, Output<IEnumerable<TSource>>> fun
+            var seed = Maybe.Of(Enumerable.Empty<TSource>());
+            Func<Maybe<IEnumerable<TSource>>, Maybe<TSource>, Maybe<IEnumerable<TSource>>> fun
                 = (m, n) => m.Bind(list => CollectCore_(n, list));
 
             return @this.Aggregate(seed, fun).AssumeNotNull();
         }
         
         // NB: We do not inline this method to avoid the creation of an unused private field (CA1823 warning).
-        private static Output<IEnumerable<TSource>> CollectCore_<TSource>(
-            Output<TSource> m, 
+        private static Maybe<IEnumerable<TSource>> CollectCore_<TSource>(
+            Maybe<TSource> m, 
             IEnumerable<TSource> list)
         {
             Contract.Requires(m != null);
 
-            return m.Bind(item => Output.Success(list.Concat(Enumerable.Repeat(item, 1))));
+            return m.Bind(item => Maybe.Of(list.Concat(Enumerable.Repeat(item, 1))));
         }
-    } // End of the class EnumerableOutputExtensions.
+
+        [SuppressMessage("Microsoft.Performance", "CA1811:AvoidUncalledPrivateCode",
+            Justification = "[GeneratedCode] This method has been overridden locally.")]
+        internal static Maybe<TSource> SumCore<TSource>(
+            this IEnumerable<Maybe<TSource>> @this)
+        {
+            Acknowledge.Object(@this);
+            Contract.Ensures(Contract.Result<Maybe<TSource>>() != null);
+
+            return @this.Aggregate(Maybe<TSource>.None, (m, n) => m.OrElse(n)).AssumeNotNull();
+        }
+    } // End of the class EnumerableMaybeExtensions.
 
     /// <content>
     /// Provides the core extension methods for <see cref="IEnumerable{T}"/>.
     /// </content>
-    internal static partial class EnumerableOutputExtensions
+    internal static partial class EnumerableMaybeExtensions
     {
         [SuppressMessage("Microsoft.Performance", "CA1811:AvoidUncalledPrivateCode",
             Justification = "[GeneratedCode] This method has been overridden locally.")]
-        internal static Output<IEnumerable<TResult>> MapCore<TSource, TResult>(
+        internal static Maybe<IEnumerable<TResult>> MapCore<TSource, TResult>(
             this IEnumerable<TSource> @this,
-            Func<TSource, Output<TResult>> funM)
+            Func<TSource, Maybe<TResult>> funM)
         {
             Acknowledge.Object(@this);
             Contract.Requires(funM != null);
-            Contract.Ensures(Contract.Result<Output<IEnumerable<TResult>>>() != null);
+            Contract.Ensures(Contract.Result<Maybe<IEnumerable<TResult>>>() != null);
 
             return @this.Select(funM).AssumeNotNull().Collect();
         }
@@ -284,7 +318,7 @@ namespace Narvalo.Fx.Advanced.Internal
             Justification = "[GeneratedCode] This method has been overridden locally.")]
         internal static IEnumerable<TSource> FilterCore<TSource>(
             this IEnumerable<TSource> @this,
-            Func<TSource, Output<bool>> predicateM)
+            Func<TSource, Maybe<bool>> predicateM)
         {
             Require.Object(@this);
             Require.NotNull(predicateM, "predicateM");
@@ -315,13 +349,14 @@ namespace Narvalo.Fx.Advanced.Internal
 
         [SuppressMessage("Microsoft.Performance", "CA1811:AvoidUncalledPrivateCode",
             Justification = "[GeneratedCode] This method has been overridden locally.")]
-        internal static Output<Tuple<IEnumerable<TFirst>, IEnumerable<TSecond>>>
+        internal static Maybe<Tuple<IEnumerable<TFirst>, IEnumerable<TSecond>>>
             MapAndUnzipCore<TSource, TFirst, TSecond>(
             this IEnumerable<TSource> @this,
-            Func<TSource, Output<Tuple<TFirst, TSecond>>> funM)
+            Func<TSource, Maybe<Tuple<TFirst, TSecond>>> funM)
         {
             Acknowledge.Object(@this);
             Contract.Requires(funM != null);
+            Contract.Ensures(Contract.Result<Maybe<Tuple<IEnumerable<TFirst>, IEnumerable<TSecond>>>>() != null);
 
             var m = @this.Select(funM).AssumeNotNull().Collect();
 
@@ -337,18 +372,18 @@ namespace Narvalo.Fx.Advanced.Internal
 
         [SuppressMessage("Microsoft.Performance", "CA1811:AvoidUncalledPrivateCode",
             Justification = "[GeneratedCode] This method has been overridden locally.")]
-        internal static Output<IEnumerable<TResult>> ZipCore<TFirst, TSecond, TResult>(
+        internal static Maybe<IEnumerable<TResult>> ZipCore<TFirst, TSecond, TResult>(
             this IEnumerable<TFirst> @this,
             IEnumerable<TSecond> second,
-            Func<TFirst, TSecond, Output<TResult>> resultSelectorM)
+            Func<TFirst, TSecond, Maybe<TResult>> resultSelectorM)
         {
             Require.NotNull(resultSelectorM, "resultSelectorM");
 
             Acknowledge.Object(@this);
             Contract.Requires(second != null);
-            Contract.Ensures(Contract.Result<Output<IEnumerable<TResult>>>() != null);
+            Contract.Ensures(Contract.Result<Maybe<IEnumerable<TResult>>>() != null);
 
-            Func<TFirst, TSecond, Output<TResult>> resultSelector
+            Func<TFirst, TSecond, Maybe<TResult>> resultSelector
                 = (v1, v2) => resultSelectorM.Invoke(v1, v2);
 
             // WARNING: Do not remove "resultSelector", otherwise .NET will make a recursive call
@@ -358,23 +393,19 @@ namespace Narvalo.Fx.Advanced.Internal
 
         [SuppressMessage("Microsoft.Performance", "CA1811:AvoidUncalledPrivateCode",
             Justification = "[GeneratedCode] This method has been overridden locally.")]
-        internal static Output<TAccumulate> FoldCore<TSource, TAccumulate>(
+        internal static Maybe<TAccumulate> FoldCore<TSource, TAccumulate>(
             this IEnumerable<TSource> @this,
             TAccumulate seed,
-            Func<TAccumulate, TSource, Output<TAccumulate>> accumulatorM)
+            Func<TAccumulate, TSource, Maybe<TAccumulate>> accumulatorM)
         {
             Require.Object(@this);
             Require.NotNull(accumulatorM, "accumulatorM");
+            Contract.Ensures(Contract.Result<Maybe<TAccumulate>>() != null);
 
-            Output<TAccumulate> result = Output.Success(seed);
+            Maybe<TAccumulate> result = Maybe.Of(seed);
 
             foreach (TSource item in @this)
             {
-                if (result == null) 
-                {
-                    return null;
-                }
-
                 result = result.Bind(_ => accumulatorM.Invoke(_, item));
             }
 
@@ -383,25 +414,27 @@ namespace Narvalo.Fx.Advanced.Internal
 
         [SuppressMessage("Microsoft.Performance", "CA1811:AvoidUncalledPrivateCode",
             Justification = "[GeneratedCode] This method has been overridden locally.")]
-        internal static Output<TAccumulate> FoldBackCore<TSource, TAccumulate>(
+        internal static Maybe<TAccumulate> FoldBackCore<TSource, TAccumulate>(
             this IEnumerable<TSource> @this,
             TAccumulate seed,
-            Func<TAccumulate, TSource, Output<TAccumulate>> accumulatorM)
+            Func<TAccumulate, TSource, Maybe<TAccumulate>> accumulatorM)
         {
             Acknowledge.Object(@this);
             Contract.Requires(accumulatorM != null);
+            Contract.Ensures(Contract.Result<Maybe<TAccumulate>>() != null);
 
             return @this.Reverse().AssumeNotNull().Fold(seed, accumulatorM);
         }
 
         [SuppressMessage("Microsoft.Performance", "CA1811:AvoidUncalledPrivateCode",
             Justification = "[GeneratedCode] This method has been overridden locally.")]
-        internal static Output<TSource> ReduceCore<TSource>(
+        internal static Maybe<TSource> ReduceCore<TSource>(
             this IEnumerable<TSource> @this,
-            Func<TSource, TSource, Output<TSource>> accumulatorM)
+            Func<TSource, TSource, Maybe<TSource>> accumulatorM)
         {
             Require.Object(@this);
             Require.NotNull(accumulatorM, "accumulatorM");
+            Contract.Ensures(Contract.Result<Maybe<TSource>>() != null);
 
             using (var iter = @this.GetEnumerator())
             {
@@ -410,15 +443,10 @@ namespace Narvalo.Fx.Advanced.Internal
                     throw new InvalidOperationException("Source sequence was empty.");
                 }
 
-                Output<TSource> result = Output.Success(iter.Current);
+                Maybe<TSource> result = Maybe.Of(iter.Current);
 
                 while (iter.MoveNext())
                 {
-                    if (result == null) 
-                    {
-                        return null;
-                    }
-
                     result = result.Bind(_ => accumulatorM.Invoke(_, iter.Current));
                 }
 
@@ -428,39 +456,36 @@ namespace Narvalo.Fx.Advanced.Internal
 
         [SuppressMessage("Microsoft.Performance", "CA1811:AvoidUncalledPrivateCode",
             Justification = "[GeneratedCode] This method has been overridden locally.")]
-        internal static Output<TSource> ReduceBackCore<TSource>(
+        internal static Maybe<TSource> ReduceBackCore<TSource>(
             this IEnumerable<TSource> @this,
-            Func<TSource, TSource, Output<TSource>> accumulatorM)
+            Func<TSource, TSource, Maybe<TSource>> accumulatorM)
         {
             Acknowledge.Object(@this);
             Contract.Requires(accumulatorM != null);
+            Contract.Ensures(Contract.Result<Maybe<TSource>>() != null);
 
             return @this.Reverse().AssumeNotNull().Reduce(accumulatorM);
         }
 
         [SuppressMessage("Microsoft.Performance", "CA1811:AvoidUncalledPrivateCode",
             Justification = "[GeneratedCode] This method has been overridden locally.")]
-        internal static Output<TAccumulate> FoldCore<TSource, TAccumulate>(
+        internal static Maybe<TAccumulate> FoldCore<TSource, TAccumulate>(
             this IEnumerable<TSource> @this,
             TAccumulate seed,
-            Func<TAccumulate, TSource, Output<TAccumulate>> accumulatorM,
-            Func<Output<TAccumulate>, bool> predicate)
+            Func<TAccumulate, TSource, Maybe<TAccumulate>> accumulatorM,
+            Func<Maybe<TAccumulate>, bool> predicate)
         {
             Require.Object(@this);
             Require.NotNull(accumulatorM, "accumulatorM");
             Require.NotNull(predicate, "predicate");
+            Contract.Ensures(Contract.Result<Maybe<TAccumulate>>() != null);
 
-            Output<TAccumulate> result = Output.Success(seed);
+            Maybe<TAccumulate> result = Maybe.Of(seed);
 
             using (var iter = @this.GetEnumerator())
             {
                 while (predicate.Invoke(result) && iter.MoveNext())
                 {
-                    if (result == null)
-                    {
-                        return null;
-                    }
-
                     result = result.Bind(_ => accumulatorM.Invoke(_, iter.Current));
                 }
             }
@@ -470,14 +495,15 @@ namespace Narvalo.Fx.Advanced.Internal
 
         [SuppressMessage("Microsoft.Performance", "CA1811:AvoidUncalledPrivateCode",
             Justification = "[GeneratedCode] This method has been overridden locally.")]
-        internal static Output<TSource> ReduceCore<TSource>(
+        internal static Maybe<TSource> ReduceCore<TSource>(
             this IEnumerable<TSource> @this,
-            Func<TSource, TSource, Output<TSource>> accumulatorM,
-            Func<Output<TSource>, bool> predicate)
+            Func<TSource, TSource, Maybe<TSource>> accumulatorM,
+            Func<Maybe<TSource>, bool> predicate)
         {
             Require.Object(@this);
             Require.NotNull(accumulatorM, "accumulatorM");
             Require.NotNull(predicate, "predicate");
+            Contract.Ensures(Contract.Result<Maybe<TSource>>() != null);
 
             using (var iter = @this.GetEnumerator())
             {
@@ -486,20 +512,15 @@ namespace Narvalo.Fx.Advanced.Internal
                     throw new InvalidOperationException("Source sequence was empty.");
                 }
 
-                Output<TSource> result = Output.Success(iter.Current);
+                Maybe<TSource> result = Maybe.Of(iter.Current);
 
                 while (predicate.Invoke(result) && iter.MoveNext())
                 {
-                    if (result == null)
-                    {
-                        return null;
-                    }
-
                     result = result.Bind(_ => accumulatorM.Invoke(_, iter.Current));
                 }
 
                 return result;
             }
         }
-    } // End of the class EnumerableOutputExtensions.
+    } // End of the class EnumerableMaybeExtensions.
 }