+++ src/Narvalo.Core/Fx/Linq/NamespaceDoc.cs
@@ -1,11 +1,12 @@
 ﻿// Copyright (c) Narvalo.Org. All rights reserved. See LICENSE.txt in the project root for license information.
 
-namespace Narvalo.Fx.Advanced
+namespace Narvalo.Fx.Linq
 {
     using System.Runtime.CompilerServices;
 
     /// <summary>
-    /// The Narvalo.Fx.Advanced namespace contains extension methods on core CLR types.
+    /// The <b>Narvalo.Fx.Linq</b> namespace provides classes and interfaces that support queries 
+    /// that use Language-Integrated Query (LINQ).
     /// </summary>
     [CompilerGenerated]
     internal static class NamespaceDoc { }