+++ src/Narvalo.Core/Fx/VoidOrError.cs
@@ -3,6 +3,7 @@
 namespace Narvalo.Fx
 {
     using System.Diagnostics;
+    using System.Diagnostics.CodeAnalysis;
     using System.Diagnostics.Contracts;
     using System.Runtime.ExceptionServices;
 
@@ -102,6 +103,8 @@ namespace Narvalo.Fx
                     _inner = inner;
                 }
 
+                [SuppressMessage("Gendarme.Rules.Performance", "AvoidUncalledPrivateCodeRule",
+                    Justification = "[Ignore] Debugger-only code.")]
                 public ExceptionDispatchInfo ExceptionInfo
                 {
                     get { return _inner._exceptionInfo; }