+++ src/Narvalo.Web/HttpQueryBinder`1.cs
@@ -10,6 +10,7 @@ namespace Narvalo.Web
     using System.Web;
 
     using Narvalo.Fx;
+    using Narvalo.Fx.Linq;
 
     public abstract partial class HttpQueryBinder<TQuery> : IHttpQueryBinder<TQuery>
     {